+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Added documentation generator to the API output.
 * Added default sort configuration for a data processor (#26).
 * Added Age field.
+* Added current user to contact filter.
 
 # Version 1.3.0